src/main/java/com/jeskert/leetcode/medium/LowestCommonAncestorOfABinaryTree.java
package com.jeskert.leetcode.medium;

import com.jeskert.leetcode.pojo.TreeNode;

public class LowestCommonAncestorOfABinaryTree {

    public TreeNode lowestCommonAncestor(TreeNode root, TreeNode p, TreeNode q) {

        while (root != null) {
            if (p.val < root.val && q.val < root.val) {
                root = root.left;
            } else if (p.val > root.val && q.val > root.val) {
                root = root.right;
            } else {
                return root;
            }
        }
        return null;
    }

}
